scripts/requirements.txt
requests>=2.31.0
yt-dlp>=2024.0.0
python-slugify>=8.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
